# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "CRUD API for managing contacts",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contacts/": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get all contacts",
        "description": "Returns every contact stored in the database.",
        "responses": {
          "200": {
            "description": "A list of contacts.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contact"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create a new contact",
        "description": "Creates a contact. All fields are required.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact created successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "6876e44e6b8f5d3e2b123456"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "One or more required fields are missing."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get a contact by ID",
        "description": "Returns a single contact matching the given ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "MongoDB ObjectId of the contact"
          }
        ],
        "responses": {
          "200": {
            "description": "The matching contact.",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "404": {
            "description": "Contact not found."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "put": {
        "tags": [
          "Contacts"
        ],
        "summary": "Update a contact by ID",
        "description": "Replaces all fields on an existing contact. All fields are required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "MongoDB ObjectId of the contact"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated contact.",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Contact not found."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete a contact by ID",
        "description": "Deletes a single contact matching the given ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "MongoDB ObjectId of the contact"
          }
        ],
        "responses": {
          "204": {
            "description": "Contact deleted successfully."
          },
          "404": {
            "description": "Contact not found."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Peter"
        },
        "lastName": {
          "type": "string",
          "example": "Parker"
        },
        "email": {
          "type": "string",
          "example": "spiderman@example.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "Black"
        },
        "birthday": {
          "type": "string",
          "example": "2001-08-10"
        }
      }
    }
  }
}